treewide: Use uthenticode for Authenticode checks
Closes #39.
